requirements.txt
## 1. requirements.txt
```
# Core ML/NLP
openai==1.3.5
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
transformers==4.36.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Web frameworks
flask==3.0.0
chainlit==1.0.0
fastapi==0.104.1
uvicorn==0.24.0

# Data processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
celery==5.3.4

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
